+++ src/org/uecide/actions/Action.java
@@ -72,5 +72,4 @@ public abstract class Action {
 
     public abstract String[] getUsage();
     public abstract String getCommand();
-
 }